refactor: extract event loop
